Fix: Resolve earnings type mismatch (Int->Double) across KoshaiProfile, fragments, adapter, and service

+++ app/src/main/java/com/example/koshailagbe/fragment/BookingDetailFragment.kt
@@ -194,7 +194,7 @@ class BookingDetailFragment : Fragment() {
                 transaction.update(bookingRef, "status", newStatus)
 
                 if (!alreadyCounted) {
-                    val totalEarnings = (currentBooking.rateBreakdown["total"] ?: 0.0).toLong()
+                    val totalEarnings = (currentBooking.rateBreakdown["total"] ?: 0.0)
                     transaction.update(koshaiRef, "totalJobs", com.google.firebase.firestore.FieldValue.increment(1))
                     transaction.update(koshaiRef, "earnings", com.google.firebase.firestore.FieldValue.increment(totalEarnings))
                     transaction.update(bookingRef, "statsCounted", true)
