+++ app/src/main/java/com/example/koshailagbe/model/KoshaiProfile.kt
@@ -14,9 +14,9 @@ data class KoshaiProfile(
     var isVerified: Boolean = false,
     var isEidMode: Boolean = false,
     var rating: Double = 0.0,
-    var totalRatings: Int = 0,
-    var totalJobs: Int = 0,
-    var earnings: Int = 0,
+    var totalRatings: Long = 0,
+    var totalJobs: Long = 0,
+    var earnings: Double = 0.0,  // Double to match Firestore FieldValue.increment()
     var ratePerCow: Double = 0.0,
     var ratePerGoat: Double = 0.0,
     var ratePerSheep: Double = 0.0
